Move bone contribution out of AutoBone
