+++ server/core/src/main/java/dev/slimevr/autobone/AutoBone.kt
@@ -7,7 +7,6 @@ import dev.slimevr.config.AutoBoneConfig
 import dev.slimevr.config.SkeletonConfig
 import dev.slimevr.poseframeformat.PoseFrameIO
 import dev.slimevr.poseframeformat.PoseFrames
-import dev.slimevr.tracking.processor.BoneType
 import dev.slimevr.tracking.processor.HumanPoseManager
 import dev.slimevr.tracking.processor.config.SkeletonConfigManager
 import dev.slimevr.tracking.processor.config.SkeletonConfigOffsets
@@ -91,75 +90,6 @@ class AutoBone(private val server: VRServer) {
 				offsets[bone] = offset
 			}
 		}
-	}
-
-	/**
-	 * Computes the local tail position of the bone after rotation.
-	 */
-	fun getBoneLocalTail(
-		skeleton: HumanPoseManager,
-		boneType: BoneType,
-	): Vector3 {
-		val bone = skeleton.getBone(boneType)
-		return bone.getTailPosition() - bone.getPosition()
-	}
-
-	/**
-	 * Computes the direction of the bone tail's movement between skeletons 1 and 2.
-	 */
-	fun getBoneLocalTailDir(
-		skeleton1: HumanPoseManager,
-		skeleton2: HumanPoseManager,
-		boneType: BoneType,
-	): Vector3? {
-		val boneOff = getBoneLocalTail(skeleton2, boneType) - getBoneLocalTail(skeleton1, boneType)
-		val boneOffLen = boneOff.len()
-		return if (boneOffLen > MIN_SLIDE_DIST) boneOff / boneOffLen else null
-	}
-
-	/**
-	 * Predicts how much the provided config should be affecting the slide offsets
-	 * of the left and right ankles.
-	 */
-	fun getSlideDot(
-		skeleton1: HumanPoseManager,
-		skeleton2: HumanPoseManager,
-		config: SkeletonConfigOffsets,
-		slideL: Vector3?,
-		slideR: Vector3?,
-	): Float {
-		var slideDot = 0f
-		// Used for right offset if not a symmetric bone
-		var boneOffL: Vector3? = null
-
-		if (slideL != null) {
-			boneOffL = getBoneLocalTailDir(skeleton1, skeleton2, config.affectedOffsets[0])
-
-			if (boneOffL != null) {
-				slideDot += slideL.dot(boneOffL)
-			}
-		}
-
-		if (slideR != null) {
-			// IMPORTANT: This assumption for acquiring BoneType only works if
-			// SkeletonConfigOffsets is set up to only affect one BoneType, make sure no
-			// changes to SkeletonConfigOffsets goes against this assumption, please!
-			val boneOffR = if (SYMM_CONFIGS.contains(config)) {
-				getBoneLocalTailDir(skeleton1, skeleton2, config.affectedOffsets[1])
-			} else if (slideL != null) {
-				// Use cached offset if slideL was used
-				boneOffL
-			} else {
-				// Compute offset if missing because of slideL
-				getBoneLocalTailDir(skeleton1, skeleton2, config.affectedOffsets[0])
-			}
-
-			if (boneOffR != null) {
-				slideDot += slideR.dot(boneOffR)
-			}
-		}
-
-		return slideDot / 2f
 	}
 
 	fun applyConfig(
@@ -538,7 +468,7 @@ class AutoBone(private val server: VRServer) {
 			val originalLength = entry.value
 
 			// Calculate the total effect of the bone based on change in rotation
-			val slideDot = getSlideDot(
+			val slideDot = BoneContribution.getSlideDot(
 				skeleton1,
 				skeleton2,
 				entry.key,
@@ -802,7 +732,7 @@ class AutoBone(private val server: VRServer) {
 
 		private fun decayFunc(initialAdjustRate: Float, adjustRateDecay: Float, epoch: Int): Float = if (epoch >= 0) initialAdjustRate / (1 + (adjustRateDecay * epoch)) else 0.0f
 
-		private val SYMM_CONFIGS = arrayOf(
+		val SYMM_CONFIGS = arrayOf(
 			SkeletonConfigOffsets.HIPS_WIDTH,
 			SkeletonConfigOffsets.SHOULDERS_WIDTH,
 			SkeletonConfigOffsets.SHOULDERS_DISTANCE,
